add adopter

+++ pet_shelter/pet_shelter/src/main/java/com/pet/pet_shelter/DTOs/Pet.java
@@ -12,7 +12,7 @@ import java.sql.Date;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Pet {
-    int petId;
+    long petId;
     String name;
     String species;
     String breed;
